requirements.txt
hikari
hikari-tanjun
requests